BUG FiX Phase 3

+++ app/src/main/java/com/mandaptak/android/Utils/Common.java
@@ -270,7 +270,7 @@ public class Common extends Application implements LayerCallbacks {
         identityProvider = new AtlasIdentityProvider(this);
     }
 
-    public String numberToWords(int number) {
+    public String numberToWords(long number) {
         if (number == 0) {
             return "zero";
         }
